+++ lib/LibDS/src/DriverStation.cpp
@@ -56,10 +56,10 @@ DriverStation::DriverStation() {
              this,      SIGNAL (communicationsChanged (DS_CommunicationStatus)));
     connect (m_manager, SIGNAL (controlModeChanged    (DS_ControlMode)),
              this,      SIGNAL (controlModeChanged    (DS_ControlMode)));
-    connect (m_manager, SIGNAL (diskUsageChanged      (int)),
-             this,      SIGNAL (diskUsageChanged      (int)));
-    connect (m_manager, SIGNAL (ramUsageChanged       (int)),
-             this,      SIGNAL (ramUsageChanged       (int)));
+    connect (m_manager, SIGNAL (diskUsageChanged      (quint8)),
+             this,      SIGNAL (diskUsageChanged      (quint8)));
+    connect (m_manager, SIGNAL (ramUsageChanged       (quint8)),
+             this,      SIGNAL (ramUsageChanged       (quint8)));
     connect (m_manager, SIGNAL (voltageChanged        (QString)),
              this,      SIGNAL (voltageChanged        (QString)));
 
@@ -217,11 +217,11 @@ void DriverStation::restartCode() {
         m_manager->protocol()->restartCode();
 }
 
-void DriverStation::startPractice (int countdown,
-                                   int autonomous,
-                                   int delay,
-                                   int teleop,
-                                   int endgame) {
+void DriverStation::startPractice (quint16 countdown,
+                                   quint16 autonomous,
+                                   quint16 delay,
+                                   quint16 teleop,
+                                   quint16 endgame) {
     /* Transform the times from seconds to milliseconds */
     delay *= 1000;
     teleop *= 1000;
@@ -230,10 +230,10 @@ void DriverStation::startPractice (int countdown,
     autonomous *= 1000;
 
     /* Calculate the time frames */
-    int _startAutonomous = countdown;
-    int _startTeleoperated = _startAutonomous + delay + autonomous;
-    int _startEndGame = _startTeleoperated + delay + teleop - endgame;
-    int _stop = _startEndGame + endgame;
+    quint32 _startAutonomous = countdown;
+    quint32 _startTeleoperated = _startAutonomous + delay + autonomous;
+    quint32 _startEndGame = _startTeleoperated + delay + teleop - endgame;
+    quint32 _stop = _startEndGame + endgame;
 
     /* Configure the sound/media timers */
     QTimer::singleShot (_startAutonomous, Qt::PreciseTimer,
@@ -290,7 +290,7 @@ void DriverStation::setProtocol (ProtocolType protocol) {
     }
 }
 
-void DriverStation::setTeamNumber (int team) {
+void DriverStation::setTeamNumber (quint16 team) {
     if (m_manager->protocolIsValid())
         m_manager->protocol()->setTeamNumber (team);
 }
@@ -314,29 +314,29 @@ void DriverStation::clearJoysticks() {
     m_manager->clearJoysticks();
 }
 
-void DriverStation::updateJoystickPovHat (int js,
-        int hat,
-        int angle) {
+void DriverStation::updateJoystickPovHat (quint8 js,
+        quint8 hat,
+        quint8 angle) {
     if (m_manager->protocolIsValid())
         m_manager->updateJoystickPovHat (js, hat, angle);
 }
 
-void DriverStation::addJoystick (int axes,
-                                 int buttons,
-                                 int povHats) {
+void DriverStation::addJoystick (quint8 axes,
+                                 quint8 buttons,
+                                 quint8 povHats) {
     if (m_manager->protocolIsValid())
         m_manager->addJoystick (axes, buttons, povHats);
 }
 
-void DriverStation::updateJoystickAxis (int js,
-                                        int axis,
+void DriverStation::updateJoystickAxis (quint8 js,
+                                        quint8 axis,
                                         double value) {
     if (m_manager->protocolIsValid())
         m_manager->updateJoystickAxis (js, axis, value);
 }
 
-void DriverStation::updateJoystickButton (int js,
-        int button,
+void DriverStation::updateJoystickButton (quint8 js,
+        quint8 button,
         bool state) {
     if (m_manager->protocolIsValid())
         m_manager->updateJoystickButton (js, button, state);
