+++ src/MainWindow/Utilities/FirstRun.h
@@ -40,7 +40,7 @@ class InitTasks {
      * If this is the first launch, it will ask the user to input his/her team
      * number using a \c QInputDialog
      */
-    static int getTeamNumber();
+    static quint16 getTeamNumber();
 
     /**
      * Initializes the CPU querying process and prompts the user to download
